+++ Nasa/src/main/java/co/Nasa/prj/promotion/service/promotion-map.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="co.Nasa.prj.promotion.service.promotion-map">
+<mapper namespace="co.Nasa.prj.promotion.service.PromotionMapper">
 </mapper>
