+++ Nasa/src/main/java/co/Nasa/prj/promotion/serviceImpl/PromotionServiceImpl.java
@@ -1,7 +1,10 @@
 package co.Nasa.prj.promotion.serviceImpl;
 
+import org.springframework.stereotype.Repository;
+
 import co.Nasa.prj.promotion.service.PromotionService;
 
+@Repository("promotionDao")
 public class PromotionServiceImpl implements PromotionService {
 
 }
